+++ src/PoEHUD/HUD/CreatureMapIcon.cs
@@ -5,7 +5,7 @@ namespace PoEHUD.HUD
 {
     public class CreatureMapIcon : MapIcon
     {
-        public CreatureMapIcon(EntityWrapper entityWrapper, string hudTexture, Func<bool> show, int iconSize) : base(entityWrapper, new HUDTexture(hudTexture), show, iconSize)
+        public CreatureMapIcon(EntityWrapper entityWrapper, string hudTexture, Func<bool> show, float iconSize) : base(entityWrapper, new HUDTexture(hudTexture), show, iconSize)
         {
         }
